#6 - Refactored matcher structure
