+++ index.ts
@@ -12,6 +12,7 @@ import {OpenCV4NodeJSVisionProvider} from "./lib/provider/opencv/opencv4nodejs.p
 import {Screen} from "./lib/screen.class";
 import {LineHelper} from "./lib/util/linehelper.class";
 
+export {jestMatchers} from "./lib/expect/jest.matcher.function";
 export {Config} from "./lib/config.class";
 export {Image} from "./lib/image.class";
 export {Key} from "./lib/key.enum";
