+++ lib/expect/matchers/toBeAt.function.ts
@@ -1,0 +1,21 @@
+import {Mouse} from "../../mouse.class";
+import {Point} from "../../point.class";
+
+export const toBeAt = async (received: Mouse, position: Point) => {
+    const currentPosition = await received.getPosition();
+
+    const success = (currentPosition.x === position.x && currentPosition.y === position.y);
+
+    if (success) {
+        return {
+            message: () =>
+                `Expected cursor to not be at position ${position.toString()}`,
+            pass: true,
+        };
+    }
+    return {
+        message: () =>
+            `Cursor should be at position ${position.toString()} but is at ${currentPosition.toString()}`,
+        pass: false,
+    };
+};
