+++ lib/expect/matchers/toShow.function.ts
@@ -1,0 +1,24 @@
+import {LocationParameters} from "../../locationparameters.class";
+import {Screen} from "../../screen.class";
+
+export const toShow = async (received: Screen, needle: string, confidence?: number) => {
+    let locationParams;
+    if (confidence) {
+        locationParams = new LocationParameters();
+        locationParams.matchProbability = confidence;
+    }
+    try {
+        await received.findOnScreen(needle, locationParams);
+        return {
+            message: () =>
+                `Expected screen to not show ${needle}`,
+            pass: true,
+        };
+    } catch (err) {
+        return {
+            message: () =>
+                `Screen is not showing ${needle}: ${err}`,
+            pass: false,
+        };
+    }
+};
